Adds workspace to heartbeat logs
Adds the workspace environment variable to the heartbeat logs.

This change allows for better identification and filtering of logs based on the specific workspace the application is running in. It also includes the public URL to the heartbeat logs.
The machine ID is renamed and used instead of the machine name.